Strim all null and empty properties from statement
The current library seems to create a lot of empty properties which
causes problems for some LRSes. This fix strims off all empty properties
before preparing the statement to send it.
Note: I have not yet had time to test this, but figured you'd want to
comment on this anyway before merging.